the *InRegion classes are now RedisDataCommands Instead of creating an instance for every function exec a stable instance is created at startup and reused for the life of the server
