+++ geode-redis/src/main/java/org/apache/geode/redis/internal/executor/key/RenameFunction.java
@@ -25,13 +25,13 @@ import org.apache.geode.cache.execute.FunctionContext;
 import org.apache.geode.cache.execute.FunctionService;
 import org.apache.geode.cache.partition.PartitionRegionHelper;
 import org.apache.geode.distributed.DistributedMember;
-import org.apache.geode.internal.cache.LocalDataSet;
 import org.apache.geode.internal.cache.PartitionedRegion;
 import org.apache.geode.internal.cache.execute.InternalFunction;
 import org.apache.geode.internal.cache.execute.RegionFunctionContextImpl;
 import org.apache.geode.redis.internal.RedisStats;
 import org.apache.geode.redis.internal.data.ByteArrayWrapper;
-import org.apache.geode.redis.internal.data.RedisKeyInRegion;
+import org.apache.geode.redis.internal.data.RedisData;
+import org.apache.geode.redis.internal.data.RedisDataCommands;
 import org.apache.geode.redis.internal.data.RedisSet;
 import org.apache.geode.redis.internal.executor.SingleResultCollector;
 import org.apache.geode.redis.internal.executor.StripedExecutor;
@@ -41,25 +41,27 @@ public class RenameFunction implements InternalFunction {
 
   public static final String ID = "REDIS_RENAME_FUNCTION";
 
-  private final transient StripedExecutor stripedExecutor;
-  private final RedisStats redisStats;
+  private final PartitionedRegion partitionedRegion;
+  private final transient RedisDataCommands dataCommands;
 
-  public static void register(StripedExecutor stripedExecutor,
+  public static void register(Region<ByteArrayWrapper, RedisData> dataRegion,
+      StripedExecutor stripedExecutor,
       RedisStats redisStats) {
-    FunctionService.registerFunction(new RenameFunction(stripedExecutor, redisStats));
+    FunctionService.registerFunction(new RenameFunction(dataRegion, stripedExecutor, redisStats));
   }
 
 
-  public RenameFunction(StripedExecutor stripedExecutor,
+  public RenameFunction(
+      Region<ByteArrayWrapper, RedisData> dataRegion,
+      StripedExecutor stripedExecutor,
       RedisStats redisStats) {
-    this.stripedExecutor = stripedExecutor;
-    this.redisStats = redisStats;
+    partitionedRegion = (PartitionedRegion) dataRegion;
+    dataCommands = new RedisDataCommands(dataRegion, redisStats, stripedExecutor);
   }
 
   @Override
   public void execute(FunctionContext context) {
     RenameContext renameContext = new RenameContext(context);
-    LocalDataSet localDataSet = (LocalDataSet) renameContext.getLocalDataSet();
 
     if (renameContext.getKeysFixedOnPrimary().size() < 2) {
       Runnable computation = () -> {
@@ -67,18 +69,11 @@ public class RenameFunction implements InternalFunction {
         context.getResultSender().lastResult(result);
       };
 
-      computeWithPrimaryLocked(renameContext.getKeyToLock(),
-          localDataSet, computation);
+      partitionedRegion.computeWithPrimaryLocked(renameContext.getKeyToLock(), computation);
     } else {
       Object result = acquireLockIfNeeded(renameContext);
       context.getResultSender().lastResult(result);
     }
-  }
-
-  public static void computeWithPrimaryLocked(Object key, LocalDataSet localDataSet, Runnable r) {
-    PartitionedRegion partitionedRegion = localDataSet.getProxy();
-
-    partitionedRegion.computeWithPrimaryLocked(key, r);
   }
 
   private boolean fixKeysOnPrimary(RenameContext context) {
@@ -97,8 +92,12 @@ public class RenameFunction implements InternalFunction {
     return getLockForNextKey(context);
   }
 
+  private StripedExecutor getStripedExecutor() {
+    return dataCommands.getStripedExecutor();
+  }
+
   private int compare(Object object1, Object object2, RenameContext context) {
-    int result = stripedExecutor.compareStripes(object1, object2);
+    int result = getStripedExecutor().compareStripes(object1, object2);
     if (result == 0) {
       DistributedMember distributedMember1 =
           PartitionRegionHelper.getPrimaryMemberForKey(context.getDataRegion(), object1);
@@ -113,7 +112,7 @@ public class RenameFunction implements InternalFunction {
 
   private boolean acquireLockIfNeeded(RenameContext context) {
     if (isLockNeededForCurrentKey(context)) {
-      return stripedExecutor
+      return getStripedExecutor()
           .execute(context.getKeyToLock(),
               () -> renameOrGetLockForNextKey(context));
     } else {
@@ -144,8 +143,7 @@ public class RenameFunction implements InternalFunction {
   }
 
   private boolean rename(RenameContext context) {
-    return new RedisKeyInRegion(context.getDataRegion(), redisStats)
-        .rename(context.getOldKey(), context.getNewKey());
+    return dataCommands.rename(context.getOldKey(), context.getNewKey());
   }
 
   private void markCurrentKeyAsLocked(RenameContext context) {
@@ -159,7 +157,7 @@ public class RenameFunction implements InternalFunction {
       ByteArrayWrapper lockedKey, RenameContext context) {
 
     boolean stripesAreTheSame =
-        stripedExecutor.compareStripes(lockedKey, context.getKeyToLock()) == 0;
+        getStripedExecutor().compareStripes(lockedKey, context.getKeyToLock()) == 0;
 
     if (!stripesAreTheSame) {
       return false;
